Pixate 2.0.0 RC 2

+++ Framework/iOS/Samples/PXShapeDemo/PXShapeDemo/PXShapeViewController.m
@@ -14,7 +14,7 @@
     CGFloat startingRotation;
     CGFloat startingScale;
     NSArray *images;
-    int imageIndex;
+    NSUInteger imageIndex;
 }
 
 - (void)loadView
